fix(kdropdown): update imports and rename files [KHCP-8879]
