+++ src/types/index.ts
@@ -8,7 +8,7 @@ export * from './catalog'
 export * from './code-block'
 export * from './collapse'
 export * from './date-time-picker'
-export * from './dropdown-menu'
+export * from './dropdown'
 export * from './modal'
 export * from './multi-select'
 export * from './table'
